# docs/swagger.yaml
basePath: /
definitions:
  main.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  main.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  main.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: Basit JWT kimlik doğrulama sistemi
  title: Fiber JWT API
  version: "1.0"
paths:
  /:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: API running
          schema:
            type: string
      summary: API durumunu döndürür
      tags:
      - Genel
  /admin:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Admin sayfasını gösterir
      tags:
      - Admin
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Giriş bilgileri
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Kullanıcı girişi yapar ve token döndürür
      tags:
      - Auth
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Giriş yapan kullanıcının bilgilerini döndürür
      tags:
      - Kullanıcı
  /refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: data
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Access token yeniler
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Kayıt bilgileri
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Yeni kullanıcı kaydı oluşturur
      tags:
      - Auth
swagger: "2.0"